chore: move application to the right package
