+++ src/commonMain/kotlin/appoutlet/gameoutlet/Main.kt
@@ -1,19 +1,17 @@
+package appoutlet.gameoutlet
+
 import androidx.compose.material.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.window.Window
-import androidx.compose.ui.window.application
+import androidx.compose.ui.window.singleWindowApplication
+import appoutlet.gameoutlet.feature.splash.SplashNavigation
+import appoutlet.gameoutlet.feature.splash.SplashScreen
 import cafe.adriel.voyager.navigator.Navigator
-import gameoutlet.feature.splash.SplashNavigation
-import gameoutlet.feature.splash.SplashScreen
-import gameoutlet.initKoin
 
 fun main() {
     val koin = initKoin()
     val defaultScreen = koin.get<SplashNavigation>().getScreen()
-    application {
-        Window(onCloseRequest = ::exitApplication) {
-            App(defaultScreen)
-        }
+    singleWindowApplication {
+        App(defaultScreen)
     }
 }
