+++ src/commonMain/kotlin/appoutlet/gameoutlet/core/CoreModules.kt
@@ -1,0 +1,11 @@
+package appoutlet.gameoutlet.core
+
+import appoutlet.gameoutlet.core.database.databaseModule
+import appoutlet.gameoutlet.core.network.networkModule
+import appoutlet.gameoutlet.core.translation.translationModule
+
+val coreModules = arrayOf(
+    translationModule,
+    databaseModule,
+    networkModule,
+)
