+++ src/commonMain/kotlin/appoutlet/gameoutlet/core/SampleClass.kt
@@ -1,4 +1,4 @@
-package gameoutlet.core
+package appoutlet.gameoutlet.core
 
 class SampleClass {
     fun sum(n1: Int, n2: Int) = n1 + n2
