+++ src/commonMain/kotlin/appoutlet/gameoutlet/core/database/DatabaseModule.kt
@@ -1,4 +1,4 @@
-package gameoutlet.core.database
+package appoutlet.gameoutlet.core.database
 
 import app.cash.sqldelight.db.SqlDriver
 import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
