+++ src/commonMain/kotlin/appoutlet/gameoutlet/feature/FeatureModules.kt
@@ -1,0 +1,7 @@
+package appoutlet.gameoutlet.feature
+
+import appoutlet.gameoutlet.feature.splash.splashModule
+
+val featureModules = arrayOf(
+    splashModule
+)
