+++ src/commonMain/kotlin/appoutlet/gameoutlet/feature/splash/SplashScreen.kt
@@ -1,4 +1,4 @@
-package gameoutlet.feature.splash
+package appoutlet.gameoutlet.feature.splash
 
 import androidx.compose.desktop.ui.tooling.preview.Preview
 import androidx.compose.foundation.layout.Box
