+++ src/commonMain/kotlin/appoutlet/gameoutlet/repository/RepositoryModules.kt
@@ -1,0 +1,7 @@
+package appoutlet.gameoutlet.repository
+
+import appoutlet.gameoutlet.repository.store.storeRepositoryModule
+
+val repositoryModules = arrayOf(
+    storeRepositoryModule,
+)
