+++ src/commonMain/kotlin/appoutlet/gameoutlet/repository/store/StoreRepository.kt
@@ -1,6 +1,6 @@
-package gameoutlet.repository.store
+package appoutlet.gameoutlet.repository.store
 
 import appoutlet.gameoutlet.core.database.GameOutletDatabase
-import gameoutlet.repository.store.api.StoreApi
+import appoutlet.gameoutlet.repository.store.api.StoreApi
 
 class StoreRepository(private val database: GameOutletDatabase,private val storeApi: StoreApi)
