+++ src/commonMain/kotlin/appoutlet/gameoutlet/repository/store/api/StoreApi.kt
@@ -1,0 +1,9 @@
+package appoutlet.gameoutlet.repository.store.api
+
+import appoutlet.gameoutlet.repository.store.api.model.StoreResponse
+import retrofit2.http.GET
+
+interface StoreApi {
+    @GET("stores")
+    suspend fun findAll(): List<StoreResponse>
+}
