+++ src/commonMain/kotlin/appoutlet/gameoutlet/repository/store/api/model/StoreResponse.kt
@@ -1,4 +1,4 @@
-package gameoutlet.repository.store.api.model
+package appoutlet.gameoutlet.repository.store.api.model
 
 data class StoreResponse(
     val storeID: String,
