--- src/commonMain/kotlin/gameoutlet/core/CoreModules.kt
@@ -1,11 +1,0 @@
-package gameoutlet.core
-
-import gameoutlet.core.database.databaseModule
-import gameoutlet.core.network.networkModule
-import gameoutlet.core.translation.translationModule
-
-val coreModules = arrayOf(
-    translationModule,
-    databaseModule,
-    networkModule,
-)
