--- src/commonMain/kotlin/gameoutlet/feature/FeatureModules.kt
@@ -1,7 +1,0 @@
-package gameoutlet.feature
-
-import gameoutlet.feature.splash.splashModule
-
-val featureModules = arrayOf(
-    splashModule
-)
