--- src/commonMain/kotlin/gameoutlet/repository/RepositoryModules.kt
@@ -1,7 +1,0 @@
-package gameoutlet.repository
-
-import gameoutlet.repository.store.storeRepositoryModule
-
-val repositoryModules = arrayOf(
-    storeRepositoryModule,
-)
